# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/online_player": {
            "get": {
                "description": "List Online Players",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "List Online Players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.OnlinePlayer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/player": {
            "get": {
                "description": "List Players",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "List Players",
                "parameters": [
                    {
                        "enum": [
                            "last_online",
                            "level"
                        ],
                        "type": "string",
                        "description": "order by field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "order by desc",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.TersePlayer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Put Players Only For SavSync,PlayerSync",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Put Players",
                "parameters": [
                    {
                        "description": "Players",
                        "name": "players",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Player"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/player/{player_uid}": {
            "get": {
                "description": "Get Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Get Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.Player"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/api/player/{player_uid}/ban": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ban Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Ban Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional ban message",
                        "name": "action",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.PlayerActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/player/{player_uid}/kick": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Kick Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Kick Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional kick message",
                        "name": "action",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.PlayerActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/player/{player_uid}/unban": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Unban Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Unban Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/whitelist": {
            "get": {
                "description": "List White List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "List White List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.PlayerW"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Put White List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Put White List",
                "parameters": [
                    {
                        "description": "Players",
                        "name": "players",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.PlayerW"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add White List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Add White List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove White List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "Remove White List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player UID",
                        "name": "player_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.EmptyResponse": {
            "type": "object"
        },
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.PlayerActionRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "database.OnlinePlayer": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "building_count": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "last_online": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "location_x": {
                    "type": "number"
                },
                "location_y": {
                    "type": "number"
                },
                "nickname": {
                    "type": "string"
                },
                "ping": {
                    "type": "number"
                },
                "player_uid": {
                    "type": "string"
                },
                "steam_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "database.Player": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "building_count": {
                    "type": "integer"
                },
                "exp": {
                    "type": "integer"
                },
                "full_stomach": {
                    "type": "number"
                },
                "hp": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "items": {
                    "$ref": "#/definitions/database.Items"
                },
                "last_online": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "location_x": {
                    "type": "number"
                },
                "location_y": {
                    "type": "number"
                },
                "max_hp": {
                    "type": "integer"
                },
                "max_status_point": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "pals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Pal"
                    }
                },
                "ping": {
                    "type": "number"
                },
                "player_uid": {
                    "type": "string"
                },
                "save_last_online": {
                    "type": "string"
                },
                "shield_hp": {
                    "type": "integer"
                },
                "shield_max_hp": {
                    "type": "integer"
                },
                "status_point": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "steam_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "database.TersePlayer": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "building_count": {
                    "type": "integer"
                },
                "exp": {
                    "type": "integer"
                },
                "full_stomach": {
                    "type": "number"
                },
                "hp": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "last_online": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "location_x": {
                    "type": "number"
                },
                "location_y": {
                    "type": "number"
                },
                "max_hp": {
                    "type": "integer"
                },
                "max_status_point": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "ping": {
                    "type": "number"
                },
                "player_uid": {
                    "type": "string"
                },
                "save_last_online": {
                    "type": "string"
                },
                "shield_hp": {
                    "type": "integer"
                },
                "shield_max_hp": {
                    "type": "integer"
                },
                "status_point": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "steam_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}